Simplify check_for_loop_arg
